trekker-backend/trekker-backend-main/requirements.txt
# Core Django
django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# CORS handling
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# Image handling
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
